TSF_fit2time

Extracts time-series data (raw or fitted) for one or many pixels and writes to an ASCII file

TSF_fit2time(
  infile,
  row_start = 1,
  row_end = Inf,
  col_start = 1,
  col_end = 1,
  outdir,
  outfile,
  wait = TRUE
)

Arguments

infile
file path of curve fitting file.

row_start
First row in processing window

row_end
Last row in processing window

col_start
First col in processing window

col_end
Last col in processing window

outdir
Output directory.

outfile
file name of output.

wait
A logical (not NA) indicating whether the R interpreter should wait for the command to finish, or run it asynchronously.

See also

Examples

if (FALSE) {
TSF_fit2time(file, 1, 1e8, 1, 1, outdir = "TSF", wait = F)
}